refactor: consolidate preferences-related functions and remove unused files
